Fix Windows vw_fs_read_file OOM and oplog unbounded malloc
src/core/vw_fs.c: (LONGLONG)SIZE_MAX overflows to -1 on 64-bit,
making sz.QuadPart > -1 always true for any non-negative file size.
Switch to unsigned comparison so the check is correct.

src/server/vw_oplog.c: payload_len / stored_plen are read from
untrusted on-disk data in seg_scan and vw_oplog_replay_from.
Add a VW_OPLOG_SEGMENT_MAX bound before malloc so a crafted segment
file cannot trigger a multi-GiB allocation.

+++ src/core/vw_fs.c
@@ -97,8 +97,8 @@ vw_err_t vw_fs_read_file(const char *path, void **out_buf, size_t *out_len) {
 
     LARGE_INTEGER sz;
     if (!GetFileSizeEx(fh, &sz)) { CloseHandle(fh); return VW_ERR_IO; }
-    if (sz.QuadPart > (LONGLONG)SIZE_MAX)  { CloseHandle(fh); return VW_ERR_OOM; }
-    if (sz.QuadPart > (LONGLONG)MAXDWORD) { CloseHandle(fh); return VW_ERR_IO; }
+    if ((ULONGLONG)sz.QuadPart > (ULONGLONG)SIZE_MAX)  { CloseHandle(fh); return VW_ERR_OOM; }
+    if ((ULONGLONG)sz.QuadPart > (ULONGLONG)MAXDWORD) { CloseHandle(fh); return VW_ERR_IO; }
 
     size_t len = (size_t)sz.QuadPart;
     void *buf = malloc(len + 1);
